packages/mock/pyproject.toml
[project]
name = "behavior-mock"
version = "1.0.0"
description = "Mock data generator service for BehaviorSense"
requires-python = ">=3.11"

dependencies = [
    "behavior-core",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pulsar-client>=3.4.0",
]

[tool.uv.sources]
behavior-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/behavior_mock"]
